+++ uploadcare_client/lib/src/api_sections/api_sections.dart
@@ -7,3 +7,4 @@ export 'files.dart';
 export 'groups.dart';
 export 'webhooks.dart';
 export 'document_converting.dart';
+export 'project.dart';