+++ uploadcare_client/lib/src/entities/entities.dart
@@ -10,3 +10,4 @@ export 'cdn.dart';
 export 'webhook.dart';
 export 'document_converting.dart';
 export 'convert.dart';
+export 'project.dart';