+++ uploadcare_client/lib/src/pubspec.dart
@@ -1,0 +1,27 @@
+// This file is generated automatically, do not modify
+
+// ignore_for_file: public_member_api_docs, constant_identifier_names, avoid_classes_with_only_static_members
+
+class Pubspec {
+  static final buildDate = DateTime.utc(2022, 5, 30, 20, 39, 33);
+
+  static const name = 'uploadcare_client';
+
+  static const versionFull = '6.0.1';
+
+  static const version = '6.0.1';
+
+  static const versionSmall = '6.0';
+
+  static const versionMajor = 6;
+
+  static const versionMinor = 0;
+
+  static const versionPatch = 1;
+
+  static const versionBuild = 0;
+
+  static const versionPreRelease = '';
+
+  static const versionIsPreRelease = false;
+}