+++ uploadcare_flutter_example/lib/main.dart
@@ -18,6 +18,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: const HomeScreen(title: 'Flutter uploadcare client example'),
+      debugShowCheckedModeBanner: false,
     );
   }
 }