+++ uploadcare_flutter_example/lib/screens/files_screen.dart
@@ -1,11 +1,15 @@
+// ignore_for_file: constant_identifier_names
+
 import 'dart:async';
 
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide AspectRatio;
 import 'package:uploadcare_flutter/uploadcare_flutter.dart';
 
 import 'face_detect_screen.dart';
 import 'file_info_screen.dart';
 import 'preview_screen.dart';
+
+enum View { ListView, CardView }
 
 class FilesScreen extends StatefulWidget {
   const FilesScreen({
@@ -24,10 +28,13 @@ class _FilesScreenState extends State<FilesScreen> {
   late int _total;
   late final StreamController<List<FileInfoEntity>> _filesController;
 
+  late View _selectedView;
+
   @override
   void initState() {
     super.initState();
 
+    _selectedView = View.ListView;
     _filesController = StreamController();
     _total = 0;
 
@@ -67,6 +74,25 @@ class _FilesScreenState extends State<FilesScreen> {
               padding: const EdgeInsets.only(right: 15.0),
               child: Center(child: Text('Total: $_total')),
             ),
+          PopupMenuButton<View>(
+              splashRadius: 25,
+              onSelected: (View item) {
+                setState(() {
+                  _selectedView = item;
+                });
+              },
+              itemBuilder: (BuildContext context) => <PopupMenuEntry<View>>[
+                    PopupMenuItem<View>(
+                      value: View.ListView,
+                      child: const Text('ListView'),
+                      enabled: _selectedView != View.ListView,
+                    ),
+                    PopupMenuItem<View>(
+                      value: View.CardView,
+                      child: const Text('CardView'),
+                      enabled: _selectedView != View.CardView,
+                    ),
+                  ]),
         ],
       ),
       body: StreamBuilder(
@@ -79,19 +105,49 @@ class _FilesScreenState extends State<FilesScreen> {
             );
           }
 
-          final files = snapshot.data!;
+          List<FileInfoEntity> files = snapshot.data!;
+
+          if (_selectedView == View.CardView) {
+            files = files.where((element) => element.isImage).toList();
+          }
 
           return ListView.builder(
             itemCount: files.length,
             itemBuilder: (context, index) {
               final file = files[index];
 
+              if (_selectedView == View.CardView) {
+                return Padding(
+                  padding: const EdgeInsets.all(10.0),
+                  child: Card(
+                    child: Padding(
+                      padding: const EdgeInsets.all(20.0),
+                      child: Image(
+                        image: UploadcareImageProvider(
+                          file.id,
+                          transformations: [
+                            ImageResizeTransformation(
+                                const Dimensions.fromHeight(1000)),
+                          ],
+                        ),
+                        fit: BoxFit.contain,
+                      ),
+                    ),
+                  ),
+                );
+              }
+
               return ListTile(
                 onTap: () => Navigator.push(
                   context,
                   MaterialPageRoute(
                     fullscreenDialog: true,
-                    builder: (context) => FileInfoScreen(file: file),
+                    builder: (context) => FileInfoScreen(
+                      file: file,
+                      api: widget.uploadcareClient.options.apiVersion < 0.7
+                          ? null
+                          : widget.uploadcareClient,
+                    ),
                   ),
                 ),
                 title: Text('Filename: ${file.filename}'),