+++ uploadcare_flutter_example/lib/screens/screens.dart
@@ -4,3 +4,4 @@ export 'face_detect_screen.dart';
 export 'file_info_screen.dart';
 export 'preview_screen.dart';
 export 'home_screen.dart';
+export 'project_info_screen.dart';