+++ uploadcare_server_mock/lib/file_routes.dart
@@ -11,7 +11,9 @@ class FileRoutes {
       ..get('/files/', _files)
       ..get('/files/<fileId>/', _fileInfo)
       ..delete('/files/storage/', _filesDelete)
-      ..get('/fakefile/<filename>', _fakeFile);
+      ..get('/fakefile/<filename>', _fakeFile)
+      ..get('/files/<fileId>/appdata/', _appData)
+      ..get('/files/<fileId>/appdata/<appName>/', _appDataByAppName);
   }
 
   final String assets;
@@ -52,5 +54,30 @@ class FileRoutes {
     final file = File(path.join(assets, filename));
     return Response.ok(file.openRead(),
         headers: {'Content-Type': 'application/octet-stream'});
+  }
+
+  Future<Response> _appData(Request request, String fileId) async {
+    final file = File(path.join(assets, 'file_application_data.json'));
+    final json = await file.readAsString();
+
+    return Response.ok(json,
+        headers: const {'Content-Type': 'application/json'});
+  }
+
+  Future<Response> _appDataByAppName(
+      Request request, String fileId, String appName) async {
+    final file = File(path.join(assets, 'file_application_data.json'));
+    final json = await file.readAsString();
+
+    Map<String, dynamic> result = jsonDecode(json);
+
+    if (!result.containsKey(appName)) {
+      return Response.notFound(appName);
+    }
+
+    result = result[appName];
+
+    return Response.ok(jsonEncode(result),
+        headers: const {'Content-Type': 'application/json'});
   }
 }